backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-aiohttp==0.1.1
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
packaging==24.2
pluggy==1.5.0
propcache==0.3.1
psycopg==3.2.6
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
websockets==15.0.1
yarl==1.19.0
